requirements.txt
amqp==5.0.6
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
prompt-toolkit==3.0.18
pytz==2021.1
redis==3.5.3
rq==1.8.0
six==1.15.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
torch==1.8.1+cpu
torchvision==0.9.1+cpu
Pillow
pandas
requests
opencv-python>=4.1.2
PyYAML>=5.3.1
matplotlib>=3.2.2
seaborn>=0.11.0
tqdm>=4.41.0
-f https://download.pytorch.org/whl/torch_stable.html